Overriding GetLetterGrade Method in RankedGradeBook Class
Overridden the GetLetterGrade Method and added some logic.
Returns a char that is the letter grade for the student.

It throws an InvalidOperationsException if there are less than 5 students.

A bunch of comparisions are made based on the threshold and grades variables within the method.

Returns 'F' character if all other comparisons failed.